+++ lib/src/utils/sura_route_navigator.dart
@@ -1,24 +1,48 @@
 import 'package:flutter/material.dart';
 
 ///Less boilerplate MaterialPageRoute builder
-class PageNavigator {
+class SuraPageNavigator {
   ///short handed push navigator
-  static Future push<T>(BuildContext context, Widget page) async {
-    return await Navigator.of(context)
-        .push<T>(MaterialPageRoute(builder: (context) => page));
+  static Future push<T>(
+    BuildContext context,
+    Widget page, {
+    RouteSettings? settings,
+  }) async {
+    return await Navigator.of(context).push<T>(
+      MaterialPageRoute(
+        builder: (context) => page,
+        settings: settings,
+      ),
+    );
   }
 
   ///short handed push replacement navigator
-  static Future pushReplacement(BuildContext context, Widget page) async {
-    return await Navigator.of(context)
-        .pushReplacement(MaterialPageRoute(builder: (context) => page));
+  static Future pushReplacement(
+    BuildContext context,
+    Widget page, {
+    RouteSettings? settings,
+  }) async {
+    return await Navigator.of(context).pushReplacement(
+      MaterialPageRoute(
+        builder: (context) => page,
+        settings: settings,
+      ),
+    );
   }
 
   ///short handed push and remove navigator
-  static Future pushAndRemove(BuildContext context, Widget page) async {
+  static Future pushAndRemove(
+    BuildContext context,
+    Widget page, {
+    RouteSettings? settings,
+    bool condition = false,
+  }) async {
     return await Navigator.of(context).pushAndRemoveUntil(
-      MaterialPageRoute(builder: (context) => page),
-      (dynamic) => false,
+      MaterialPageRoute(
+        builder: (context) => page,
+        settings: settings,
+      ),
+      (dynamic) => condition,
     );
   }